% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circlepatch_class.R
\name{circlepatch}
\alias{circlepatch}
\alias{plot.circlepatch}
\alias{print.circlepatch}
\alias{summary.circlepatch}
\title{circlepatch class description}
\usage{
\method{print}{circlepatch}(x)

\method{plot}{circlepatch}(x, border = "black", col = "red",
  background = rgb(1, 1, 1, alpha = 1), add = F, main = NULL)

\method{summary}{circlepatch}(x)
}
\arguments{
\item{x}{a \code{circlepatch} object generated for instance by
\code{\link{generate_circle_patch}}}
}
\description{
circlepatch class description

Print object of \code{circlepatch} class

Plot object of \code{circlepatch} class

Summary method for \code{circlepatch} object
}

