% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration_class.R
\name{exploration}
\alias{exploration}
\alias{plot.exploration}
\alias{print.exploration}
\alias{summary.exploration}
\title{exploration class description}
\usage{
\method{print}{exploration}(x)

\method{plot}{exploration}(x, type = "marginal", wait = T)

\method{summary}{exploration}(x)
}
\arguments{
\item{x}{a \code{exploration} object generated for instance by
\code{\link{explore}}}
}
\description{
exploration class description

Print object of \code{exploration} class

Plot object of \code{exploration} class

Summary of \code{exploration} class
}

