% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_exploration.R
\name{explore}
\alias{explore}
\title{Analyse exploration around a central place like a waterhole}
\usage{
explore(data, buffer = 100, maxdist = 10000, binsize = 1000, projstring,
  waterholes)
}
\arguments{
\item{data}{a data frame with coordinates  \code{data$x} and \code{data$y},
and nearest waterhole in column \code{data$name}. It should be a movement
sequence for a single animal and a single exploration event.}

\item{buffer}{width of exploration around the trajectory.}

\item{maxdist}{maximum distance to the central place above which we ignore
exploration}

\item{binsize}{width of distance rings}

\item{projstring}{CRS projection of data}

\item{waterholes}{SpatialPointsDataFrame with waterholes location and column
\code{$name} in data slot.}
}
\description{
\code{explore} compute percentage and total area explored in each distance
class around the central place, for each day. It also compute the new area
explored at the beginning of the first day. It can work with any sequence of
movement, given each sequence as a unique \code{index} value
}
\examples{
explore(data,buffer=100,maxdist=10000,binsize=1000,projstring,waterholes)
}

