% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_exploration.R
\name{explore_one_night}
\alias{explore_one_night}
\title{Analyse exploration around a central place for a single night}
\usage{
explore_one_night(night, central_point, already_explored, buffer = 100,
  projstring, marginal = F)
}
\arguments{
\item{night}{a data frame with coordinates  \code{data$x} and \code{data$y}}

\item{central_point}{single feature SpatialPointsDataFrame with column
\code{$name} in data slot.}

\item{already_explored}{\code{\link{circlepatch}} object containing areas
unexplored and already explored}

\item{buffer}{width of exploration around the trajectory.}
}
\description{
\code{explore_one_night} compute percentage and total area explored in each
distance class around the central place, for each day. It also compute the
new area explored at the beginning of the first day. It can work with any
sequence of movement, given each sequence as a unique \code{index} value
}
\examples{
explore_one_night(night,central_point,already_explored,buffer=100,projstring,marginal=F)
}

