% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_patch_function.R
\name{generate_circle_patch}
\alias{generate_circle_patch}
\title{Divide a circular disk around a central point into concentric rings.}
\usage{
generate_circle_patch(central_point, maxdist = 10000, binsize = 1000,
  nptsPerimeter = 100, projstring)
}
\arguments{
\item{central_point}{a SpatialPoints or SpatialPointsDataFrame in the center of the patch.}

\item{maxdist}{maximum distance to the central place above which we ignore
exploration}

\item{binsize}{width of distance rings}

\item{nptsPerimeter}{number of points in perimeter}

\item{projstring}{CRS projection of data}
}
\description{
Divide a circular disk around a central point into concentric rings.
}

