% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatchExplo.R
\docType{package}
\name{PatchExplo}
\alias{PatchExplo}
\alias{PatchExplo-package}
\title{PatchExplo: Describing patch exploration}
\description{
PatchExplo provides usefull tools to describe patterns of patch exploration.
For now it works only on patch organized around central points i.e. a waterhole
and describe exploration on successive day for each distance to water.
}

