% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circlepatch_class.R
\name{plot.summary.circlepatch}
\alias{plot.summary.circlepatch}
\title{Plot method for \code{summary.circlepatch} object}
\usage{
\method{plot}{summary.circlepatch}(x, xlab = "Distance to waterhole",
  ylab = "Explored Area", ncol = 2)
}
\arguments{
\item{x}{a \code{summary.circlepatch} object}
}
\description{
Used to build its summary
}

