% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration_class.R
\name{plot.summary.exploration}
\alias{plot.summary.exploration}
\title{Plot method for \code{summary.exploration} object}
\usage{
\method{plot}{summary.exploration}(x, xlab = "Distance to waterhole",
  ylab = "Explored Area", ncol = 2, current = T, marginal = T,
  wait = T, title = NULL)
}
\arguments{
\item{x}{a \code{summary.exploration} object}
}
\description{
Used to build its summary
}

