% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot_waterholes}
\alias{plot_waterholes}
\title{Plot waterholes with names labeled}
\usage{
plot_waterholes(x)
}
\arguments{
\item{x}{SpatialPointsDataFrame}
}
\description{
Plot waterholes with names labeled
}

