% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{prevent_invalid_geometry}
\alias{prevent_invalid_geometry}
\title{Prevent geometry from being invalid}
\usage{
prevent_invalid_geometry(x)
}
\arguments{
\item{x}{SpatialPolygons or SpatialPolygonsDataFrame}
}
\description{
Prevent geometry from being invalid
}

