% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circlepatch_class.R
\name{print.summary.circlepatch}
\alias{print.summary.circlepatch}
\title{Print method for \code{summary.circlepatch} object}
\usage{
\method{print}{summary.circlepatch}(x)
}
\arguments{
\item{x}{a \code{summary.circlepatch} object}
}
\description{
Print method for \code{summary.circlepatch} object
}

