% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration_class.R
\name{print.summary.exploration}
\alias{print.summary.exploration}
\title{Print method for \code{summary.exploration} object}
\usage{
\method{print}{summary.exploration}(x)
}
\arguments{
\item{x}{a \code{summary.exploration} object}
}
\description{
Print output
}

