% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{zone2data}
\alias{zone2data}
\title{Analyze a concentric ring of a \code{circlepatch} object}
\usage{
zone2data(zone)
}
\arguments{
\item{zone}{one element of a \code{circlepatch} object}
}
\description{
Used to build its summary
}

